Preparation for the BAUHAUS release. Updated buildpack to work with correctly formed seven5 apps.
Detection is now "seven5" and it vaguely checks that the project is laid out right.
Not clear that it will work if you have several user defined packages in your dart heirarchy.
The symlinks produced by "pub" cause a horrible mess on heroku and had to mv the lib directory directly
into the packages primary directory to get 'packages/appname' working ok. Horrible.